# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mini Project tentang penyewaan alat-alat bertema lingkungan mulai dari alat perawatan tanaman, kebersihan lingkungan, alat hiking dan camping, dan alat proses daur ulang.",
        "title": "Chandra Mini Project",
        "contact": {
            "url": "http://www.swagger.io/support"
        },
        "version": "1.0"
    },
    "basePath": "/v2",
    "paths": {
...
        "/admin/equipment": {
            "post": {
                "description": "Post Equipment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment"
                ],
                "summary": "Post Equipment",
                "operationId": "Post-Equipment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.EquipmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/equipment/category": {
            "post": {
                "description": "Post Category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Post Category",
                "operationId": "Post-Category",
                "responses": {
                    "200": {
                        "description": "Create data Success",
                        "schema": {
                            "$ref": "#/definitions/response.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/equipment": {
            "get": {
                "description": "Get All Equipment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment"
                ],
                "summary": "Get All Equipment",
                "operationId": "Get-All-Equipment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.EquipmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/equipment/category": {
            "get": {
                "description": "Get All Category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Get All Category",
                "operationId": "Get-All-Category",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/rent": {
            "post": {
                "description": "Get All Data Rent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Get All Data Rent",
                "operationId": "Post-Rent",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/rent/user": {
            "get": {
                "description": "Get All Data Rent For User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Get All Data Rent For User",
                "operationId": "Get-User-Rent",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/rent/user{id}": {
            "put": {
                "description": "Update Data Rent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Update Data Rent",
                "operationId": "Update-Rent",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "prototype_api_controller_rent_response.UserData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "response.CategoryDetails": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Daur Ulang"
                }
            }
        },
        "response.EquipmentDetails": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "response.EquipmentResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "CategoryId  int             `json:\"category_id\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.CategoryDetails"
                        }
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "Mesin Daur Ulang Kecil 1.5 Ton"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "https://cloudinary.com/photo/2016/03/31/15/32/robot-1295393_960_720.png"
                },
                "name": {
                    "type": "string",
                    "example": "Mesin Daur Ulang Kecil 1.5 Ton"
                },
                "price": {
                    "type": "integer",
                    "example": 1000000
                },
                "stock": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "response.RentResponse": {
            "type": "object",
            "properties": {
                "equipment": {
                    "description": "EquipmentId int              `json:\"equipment_id\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.EquipmentDetails"
                        }
                    ]
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100000
                },
                "user_data": {
                    "description": "UserId uuid.UUID `json:\"user_id\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/prototype_api_controller_rent_response.UserData"
                        }
                    ]
                }
            }
        },
...
    }
}